services/ingestion/requirements.txt
httpx==0.27.2
kafka-python==2.0.2
pydantic==2.9.2
psycopg[binary]==3.2.3
websockets==13.1
